Customize sonar a little.
We want the Python stuff installed so we might as well extend the base
docker container and add extensions.  The sonar installation is also
exposed through the frontend and bound to port 9000 on the host.  The
latter is important since pipeline containers do not have access to the
docker mappings (e.g., those in /etc/hosts) :/